[android][drape_frontend] Make the unity build batch size a speed knob
Several files under android/sdk/src/main/cpp used identifiers they never
declared. They compiled only because a sibling .cpp happened to precede them
in the same unity build translation unit and leaked the name, which made
CMAKE_UNITY_BUILD_BATCH_SIZE decide whether the code compiles rather than how
fast it compiles.

Fix the consumers -- missing declarations, includes and qualification -- and
remove the leak sources: no namespace-scope using-directive or
using-declaration is left in android/sdk. An anonymous namespace is not a
shield, because the implicit using-directive that exposes its contents at file
scope exposes its own using-directives there too.

Merge the routing JNI helpers into RoutingJni.cpp/hpp, which also makes 11
previously exported symbols file-local.

libs/drape_frontend/arrow3d.cpp includes 3party/fast_obj/fast_obj.h, which
defines unprefixed macros (array_size, BUFFER_SIZE, ...) and never undefines
them, poisoning the rest of its blob. Exclude it from unity builds, as
libs/platform and generator already do for the same reason.

Add CMAKE_UNITY_BUILD_BATCH_SIZE and Gradle -Punity / -PunityBatchSize hooks
so any batch size can be verified.
