+++ android/sdk/build.gradle
@@ -40,6 +40,14 @@ android {
             '-DBUILD_TESTING=OFF', '-DSKIP_TOOLS=ON', "-DUSE_PCH=$pchFlag",
             "-DNJOBS=$njobs", "-DENABLE_VULKAN_DIAGNOSTICS=$enableVulkanDiagnostics",
             "-DENABLE_TRACE=$enableTrace"
+        // Passed through only when set, so CMake keeps owning the defaults.
+        // Use -Punity=OFF or -PunityBatchSize=3 to check that the code does not depend on the batching.
+        if (project.hasProperty('unity')) {
+          arguments "-DCMAKE_UNITY_BUILD=${project.getProperty('unity')}"
+        }
+        if (project.hasProperty('unityBatchSize')) {
+          arguments "-DCMAKE_UNITY_BUILD_BATCH_SIZE=${project.getProperty('unityBatchSize')}"
+        }
         targets 'organicmaps'
       }
     }
