+++ android/sdk/src/main/cpp/app/organicmaps/sdk/DownloadResourcesLegacyActivity.cpp
@@ -24,11 +24,6 @@
 #include <string>
 #include <vector>
 
-using namespace downloader;
-using namespace storage;
-
-using namespace std::placeholders;
-
 /// Special error codes to notify GUI about free space
 //@{
 #define ERR_DOWNLOAD_SUCCESS      0
@@ -44,28 +39,14 @@ namespace
 {
 
 static std::vector<platform::CountryFile> g_filesToDownload;
-static int g_totalDownloadedBytes;
-static int g_totalBytesToDownload;
-static std::shared_ptr<HttpRequest> g_currentRequest;
+static uint64_t g_totalDownloadedBytes;
+static uint64_t g_totalBytesToDownload;
+static std::shared_ptr<downloader::HttpRequest> g_currentRequest;
 
 }  // namespace
 
 extern "C"
 {
-using Callback = HttpRequest::Callback;
-
-static int HasSpaceForFiles(Platform & pl, std::string const & sdcardPath, size_t fileSize)
-{
-  switch (pl.GetWritableStorageStatus(fileSize))
-  {
-  case Platform::STORAGE_DISCONNECTED: return ERR_STORAGE_DISCONNECTED;
-
-  case Platform::NOT_ENOUGH_SPACE: return ERR_NOT_ENOUGH_FREE_SPACE;
-
-  default: return static_cast<int>(fileSize);
-  }
-}
-
 JNIEXPORT jint Java_app_organicmaps_sdk_DownloadResourcesLegacyActivity_nativeGetBytesToDownload(JNIEnv * env,
                                                                                                  jclass clazz)
 {
@@ -74,32 +55,39 @@ JNIEXPORT jint Java_app_organicmaps_sdk_DownloadResourcesLegacyActivity_nativeGe
   g_totalBytesToDownload = 0;
   g_totalDownloadedBytes = 0;
 
-  using namespace storage;
-  Storage const & storage = g_framework->GetStorage();
+  auto const & storage = g_framework->GetStorage();
   auto const status = storage.GetForceDownloadWorlds(g_filesToDownload);
 
   for (auto const & cf : g_filesToDownload)
     g_totalBytesToDownload += cf.GetRemoteSize();
 
-  int res;
-  if (status == Storage::WorldStatus::ERROR_CREATE_FOLDER || status == Storage::WorldStatus::ERROR_MOVE_FILE)
-  {
+  // Only World and WorldCoasts are downloaded here, so the total always fits into jint.
+  jint res;
+  using enum storage::Storage::WorldStatus;
+  if (status == ERROR_CREATE_FOLDER || status == ERROR_MOVE_FILE)
     res = ERR_DISK_ERROR;
-  }
   else
   {
-    Platform & pl = GetPlatform();
-    res = HasSpaceForFiles(pl, pl.WritableDir(), g_totalBytesToDownload);
-  }
+    auto const & pl = GetPlatform();
+    switch (pl.GetWritableStorageStatus(g_totalBytesToDownload))
+    {
+    case Platform::STORAGE_DISCONNECTED:
+      res = ERR_STORAGE_DISCONNECTED;
+      LOG(LWARNING, ("External file system is not available"));
+      break;
 
-  if (res == ERR_STORAGE_DISCONNECTED)
-    LOG(LWARNING, ("External file system is not available"));
-  else if (res == ERR_NOT_ENOUGH_FREE_SPACE)
-    LOG(LWARNING, ("Not enough space to extract files"));
+    case Platform::NOT_ENOUGH_SPACE:
+      res = ERR_NOT_ENOUGH_FREE_SPACE;
+      LOG(LWARNING, ("Not enough space to download files of", g_totalBytesToDownload, "bytes"));
+      break;
+
+    default: res = static_cast<jint>(g_totalBytesToDownload);
+    }
+  }
 
   g_currentRequest.reset();
 
-  if (status == Storage::WorldStatus::WAS_MOVED)
+  if (status == WAS_MOVED)
   {
     g_framework->ReloadWorldMaps();
     res = ERR_DOWNLOAD_SUCCESS;  // reset possible storage error if we moved files
@@ -108,13 +96,13 @@ JNIEXPORT jint Java_app_organicmaps_sdk_DownloadResourcesLegacyActivity_nativeGe
   return res;
 }
 
-static void DownloadFileFinished(std::shared_ptr<jobject> obj, HttpRequest const & req)
+static void DownloadFileFinished(std::shared_ptr<jobject> obj, downloader::HttpRequest const & req)
 {
   auto const status = req.GetStatus();
-  ASSERT_NOT_EQUAL(status, DownloadStatus::InProgress, ());
+  ASSERT_NOT_EQUAL(status, downloader::DownloadStatus::InProgress, ());
 
   int errorCode = ERR_DOWNLOAD_ERROR;
-  if (status == DownloadStatus::Completed)
+  if (status == downloader::DownloadStatus::Completed)
     errorCode = ERR_DOWNLOAD_SUCCESS;
 
   g_currentRequest.reset();
@@ -137,7 +125,7 @@ static void DownloadFileFinished(std::shared_ptr<jobject> obj, HttpRequest const
   env->CallVoidMethod(*obj, methodID, errorCode);
 }
 
-static void DownloadFileProgress(std::shared_ptr<jobject> listener, HttpRequest const & req)
+static void DownloadFileProgress(std::shared_ptr<jobject> listener, downloader::HttpRequest const & req)
 {
   JNIEnv * env = jni::GetEnv();
   static jmethodID methodID = jni::GetMethodID(env, *listener, "onProgress", "(I)V");
@@ -149,12 +137,14 @@ JNIEXPORT jint Java_app_organicmaps_sdk_DownloadResourcesLegacyActivity_nativeSt
                                                                                                     jclass clazz,
                                                                                                     jobject listener)
 {
+  using namespace std::placeholders;
+
   if (g_filesToDownload.empty())
     return ERR_NO_MORE_FILES;
 
   /// @todo One downloader instance with cached servers. All this routine will be refactored some time.
   static auto downloader = storage::GetDownloader();
-  storage::Storage const & storage = g_framework->GetStorage();
+  auto const & storage = g_framework->GetStorage();
   downloader->SetDataVersion(storage.GetCurrentDataVersion());
 
   downloader->EnsureMetaConfigReady([&storage, ptr = jni::make_global_ref(listener)]()
@@ -163,10 +153,10 @@ JNIEXPORT jint Java_app_organicmaps_sdk_DownloadResourcesLegacyActivity_nativeSt
     auto const fileName = curFile.GetFileName(MapFileType::Map);
     LOG(LINFO, ("Downloading file", fileName));
 
-    g_currentRequest.reset(HttpRequest::GetFile(downloader->MakeUrlListLegacy(fileName),
-                                                storage.GetFilePath(curFile.GetName(), MapFileType::Map),
-                                                curFile.GetRemoteSize(), std::bind(&DownloadFileFinished, ptr, _1),
-                                                std::bind(&DownloadFileProgress, ptr, _1), false));
+    g_currentRequest.reset(downloader::HttpRequest::GetFile(
+        downloader->MakeUrlListLegacy(fileName), storage.GetFilePath(curFile.GetName(), MapFileType::Map),
+        curFile.GetRemoteSize(), std::bind(&DownloadFileFinished, ptr, _1), std::bind(&DownloadFileProgress, ptr, _1),
+        false));
   });
 
   return ERR_FILE_IN_PROGRESS;
