+++ android/sdk/src/main/cpp/app/organicmaps/sdk/Framework.hpp
@@ -192,11 +192,6 @@ public:
   void OnPowerFacilityChanged(power_management::Facility const facility, bool enabled) override;
   void OnPowerSchemeChanged(power_management::Scheme const actualScheme) override;
 };
-
-namespace framework
-{
-jint registerNativeMethods(JNIEnv * env);
-}  // namespace framework
 }  // namespace android
 
 extern CheckedPtr<android::Framework> g_framework;
