+++ android/sdk/src/main/cpp/app/organicmaps/sdk/FrameworkJni.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <jni.h>
+
+namespace android::framework
+{
+jint registerNativeMethods(JNIEnv * env);
+}  // namespace android::framework
