+++ android/sdk/src/main/cpp/app/organicmaps/sdk/TrackRecorder.cpp
@@ -1,6 +1,8 @@
 #include "Framework.hpp"
 
+#include "app/organicmaps/sdk/bookmarks/data/ElevationInfo.hpp"
 #include "app/organicmaps/sdk/core/jni_helper.hpp"
+
 #include "map/gps_tracker.hpp"
 
 #include <chrono>
@@ -26,7 +28,8 @@ JNIEXPORT void Java_app_organicmaps_sdk_location_TrackRecorder_nativeSetTrackRec
   frm()->SetTrackRecordingUpdateHandler(
       [listener = jni::make_global_ref(updateListener)](TrackStatistics const & trackStats)
   {
-    JNIEnv * env = jni::GetEnvSafe();
+    // The handler is a platform::SafeCallback, so it always runs on the attached GUI thread.
+    JNIEnv * env = jni::GetEnv();
     jobject stats =
         env->NewObject(g_trackStatisticsClazz, cId, trackStats.m_length, trackStats.m_duration, trackStats.m_ascent,
                        trackStats.m_descent, trackStats.m_minElevation, static_cast<jint>(trackStats.m_maxElevation));
