+++ android/sdk/src/main/cpp/app/organicmaps/sdk/bookmarks/data/Bookmark.hpp
@@ -2,6 +2,8 @@
 
 #include <jni.h>
 
+#include "app/organicmaps/sdk/core/jni_helper.hpp"
+
 #include "map/bookmark.hpp"
 
 jobject CreateBookmark(JNIEnv * env, place_page::Info const & info, jni::TScopedLocalObjectArrayRef const & jrawTypes,
