+++ android/sdk/src/main/cpp/app/organicmaps/sdk/bookmarks/data/BookmarkManager.cpp
@@ -18,11 +18,9 @@
 #include "base/macros.hpp"
 #include "base/string_utils.hpp"
 
+#include <functional>
 #include <limits>
 #include <utility>
-
-using namespace jni;
-using namespace std::placeholders;
 
 namespace
 {
@@ -241,6 +239,8 @@ Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeShowBookmarkCatego
 
 JNIEXPORT void Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeLoadBookmarks(JNIEnv * env, jclass)
 {
+  using namespace std::placeholders;
+
   PrepareClassRefs(env);
   BookmarkManager::AsyncLoadingCallbacks callbacks;
   callbacks.m_onStarted = std::bind(&OnAsyncLoadingStarted, env);
@@ -257,7 +257,7 @@ JNIEXPORT void Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeLoa
 JNIEXPORT jlong Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeCreateCategory(JNIEnv * env, jobject,
                                                                                              jstring name)
 {
-  auto const categoryId = frm()->GetBookmarkManager().CreateBookmarkCategory(ToNativeString(env, name));
+  auto const categoryId = frm()->GetBookmarkManager().CreateBookmarkCategory(jni::ToNativeString(env, name));
   frm()->GetBookmarkManager().SetLastEditedBmCategory(categoryId);
   return static_cast<jlong>(categoryId);
 }
@@ -325,7 +325,7 @@ JNIEXPORT void Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeLoa
                                                                                                jstring path,
                                                                                                jboolean isTemporaryFile)
 {
-  frm()->AddBookmarksFile(ToNativeString(env, path), isTemporaryFile);
+  frm()->AddBookmarksFile(jni::ToNativeString(env, path), isTemporaryFile);
 }
 
 JNIEXPORT jboolean JNICALL
@@ -378,11 +378,6 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_native
                         address);
 }
 
-static uint32_t shift(uint32_t v, uint8_t bitCount)
-{
-  return v << bitCount;
-}
-
 JNIEXPORT jobject Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeGetTrack(JNIEnv * env, jobject,
                                                                                          jlong trackId,
                                                                                          jclass trackClazz)
@@ -405,7 +400,7 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_native
 JNIEXPORT jboolean JNICALL
 Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeIsUsedCategoryName(JNIEnv * env, jclass, jstring name)
 {
-  return static_cast<jboolean>(frm()->GetBookmarkManager().IsUsedCategoryName(ToNativeString(env, name)));
+  return static_cast<jboolean>(frm()->GetBookmarkManager().IsUsedCategoryName(jni::ToNativeString(env, name)));
 }
 
 JNIEXPORT void Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativePrepareForSearch(JNIEnv *, jclass,
@@ -510,14 +505,10 @@ JNIEXPORT void Java_app_organicmaps_sdk_bookmarks_data_BookmarkManager_nativeGet
   sortParams.m_sortingType = static_cast<BookmarkManager::SortingType>(sortingType);
   sortParams.m_hasMyPosition = static_cast<bool>(hasMyPosition);
   sortParams.m_myPosition = mercator::FromLatLon(static_cast<double>(lat), static_cast<double>(lon));
-  sortParams.m_onResults = bind(&OnCategorySortingResults, env, timestamp, _1, _2);
+  sortParams.m_onResults =
+      std::bind(&OnCategorySortingResults, env, timestamp, std::placeholders::_1, std::placeholders::_2);
 
   bm.GetSortedCategory(sortParams);
-}
-
-constexpr static uint8_t ExtractByte(uint32_t number, uint8_t byteIdx)
-{
-  return (number >> (8 * byteIdx)) & 0xFF;
 }
 
 JNIEXPORT void JNICALL
