+++ android/sdk/src/main/cpp/app/organicmaps/sdk/core/logging.cpp
@@ -13,10 +13,9 @@
 
 namespace jni
 {
-
-using namespace base;
-
-void AndroidMessage(LogLevel level, SrcPoint const & src, std::string const & s)
+namespace
+{
+void AndroidMessage(base::LogLevel level, base::SrcPoint const & src, std::string const & s)
 {
   android_LogPriority pr = ANDROID_LOG_SILENT;
 
@@ -39,33 +38,34 @@ void AndroidMessage(LogLevel level, SrcPoint const & src, std::string const & s)
   env->CallStaticVoidMethod(g_loggerClazz, logMethod, pr, NULL, msg.get(), NULL);
 }
 
-void AndroidLogMessage(LogLevel level, SrcPoint const & src, std::string const & s)
+void AndroidLogMessage(base::LogLevel level, base::SrcPoint const & src, std::string const & s)
 {
   AndroidMessage(level, src, s);
-  CHECK_LESS(level, g_LogAbortLevel, ("Abort. Log level is too serious", level));
+  CHECK_LESS(level, base::g_LogAbortLevel, ("Abort. Log level is too serious", level));
 }
 
-bool AndroidAssertMessage(SrcPoint const & src, std::string const & s)
+bool AndroidAssertMessage(base::SrcPoint const & src, std::string const & s)
 {
   AndroidMessage(LCRITICAL, src, s);
   return true;
 }
+}  // namespace
 
 void InitSystemLog()
 {
-  SetLogMessageFn(&AndroidLogMessage);
+  base::SetLogMessageFn(&AndroidLogMessage);
 }
 
 void InitAssertLog()
 {
-  SetAssertFunction(&AndroidAssertMessage);
+  base::SetAssertFunction(&AndroidAssertMessage);
 }
 
 void ToggleDebugLogs(bool enabled)
 {
   if (enabled)
-    g_LogLevel = LDEBUG;
+    base::g_LogLevel = LDEBUG;
   else
-    g_LogLevel = LINFO;
+    base::g_LogLevel = LINFO;
 }
 }  // namespace jni
