+++ android/sdk/src/main/cpp/app/organicmaps/sdk/editor/OpeningHours.cpp
@@ -16,12 +16,6 @@
 
 namespace
 {
-using namespace editor;
-using namespace editor::ui;
-using namespace osmoh;
-using THours = std::chrono::hours;
-using TMinutes = std::chrono::minutes;
-
 // ID-s for HoursMinutes class
 jclass g_clazzHoursMinutes;
 jmethodID g_ctorHoursMinutes;
@@ -77,14 +71,14 @@ jobject JavaTimetable(JNIEnv * env, jobject workingHours, jobject closedHours, b
   return tt;
 }
 
-jobject JavaTimetable(JNIEnv * env, TimeTable const & tt)
+jobject JavaTimetable(JNIEnv * env, editor::ui::TimeTable const & tt)
 {
   auto const excludeSpans = tt.GetExcludeTime();
-  std::set<Weekday> weekdays = tt.GetOpeningDays();
+  std::set<osmoh::Weekday> weekdays = tt.GetOpeningDays();
   std::vector<int> weekdaysVector;
   weekdaysVector.reserve(weekdays.size());
   std::transform(weekdays.begin(), weekdays.end(), std::back_inserter(weekdaysVector),
-                 [](Weekday weekday) { return static_cast<int>(weekday); });
+                 [](osmoh::Weekday weekday) { return static_cast<int>(weekday); });
   jintArray jWeekdays = env->NewIntArray(static_cast<jsize>(weekdays.size()));
   env->SetIntArrayRegion(jWeekdays, 0, static_cast<jsize>(weekdaysVector.size()), &weekdaysVector[0]);
 
@@ -95,7 +89,7 @@ jobject JavaTimetable(JNIEnv * env, TimeTable const & tt)
       tt.IsTwentyFourHours(), jWeekdays);
 }
 
-jobjectArray JavaTimetables(JNIEnv * env, TimeTableSet & tts)
+jobjectArray JavaTimetables(JNIEnv * env, editor::ui::TimeTableSet & tts)
 {
   size_t const size = tts.Size();
   jobjectArray const result = env->NewObjectArray(static_cast<jsize>(size), g_clazzTimetable, 0);
@@ -108,10 +102,10 @@ jobjectArray JavaTimetables(JNIEnv * env, TimeTableSet & tts)
   return result;
 }
 
-jobject JavaOpeningHoursInfo(JNIEnv * env, RuleState state, bool isTwentyFourSeven, time_t nextTimeOpen,
+jobject JavaOpeningHoursInfo(JNIEnv * env, osmoh::RuleState state, bool isTwentyFourSeven, time_t nextTimeOpen,
                              time_t nextTimeClosed)
 {
-  ASSERT(state != RuleState::Unknown, ("Shouldn't instantiate java OpeningHours with unknown state"));
+  ASSERT(state != osmoh::RuleState::Unknown, ("Shouldn't instantiate java OpeningHours with unknown state"));
 
   jlong javaTimeNeverConstant = static_cast<jlong>(-1);
   jlong jlongNextTimeOpen = static_cast<jlong>(nextTimeOpen);
@@ -128,30 +122,30 @@ jobject JavaOpeningHoursInfo(JNIEnv * env, RuleState state, bool isTwentyFourSev
   return info;
 }
 
-HourMinutes NativeHoursMinutes(JNIEnv * env, jobject jHourMinutes)
+osmoh::HourMinutes NativeHoursMinutes(JNIEnv * env, jobject jHourMinutes)
 {
   jlong const hours = env->GetLongField(jHourMinutes, g_fidHours);
   jlong const minutes = env->GetLongField(jHourMinutes, g_fidMinutes);
-  return HourMinutes(THours(hours) + TMinutes(minutes));
+  return osmoh::HourMinutes(std::chrono::hours(hours) + std::chrono::minutes(minutes));
 }
 
-Timespan NativeTimespan(JNIEnv * env, jobject jTimespan)
+osmoh::Timespan NativeTimespan(JNIEnv * env, jobject jTimespan)
 {
-  Timespan span;
+  osmoh::Timespan span;
   span.SetStart(NativeHoursMinutes(env, env->GetObjectField(jTimespan, g_fidStart)));
   span.SetEnd(NativeHoursMinutes(env, env->GetObjectField(jTimespan, g_fidEnd)));
   return span;
 }
 
-TimeTable NativeTimetable(JNIEnv * env, jobject jTimetable)
+editor::ui::TimeTable NativeTimetable(JNIEnv * env, jobject jTimetable)
 {
-  TimeTable tt = TimeTable::GetPredefinedTimeTable();
+  auto tt = editor::ui::TimeTable::GetPredefinedTimeTable();
   jintArray const jWeekdays = static_cast<jintArray>(env->GetObjectField(jTimetable, g_fidWeekdays));
   int * weekdaysArr = static_cast<int *>(env->GetIntArrayElements(jWeekdays, nullptr));
   jint size = env->GetArrayLength(jWeekdays);
-  std::set<Weekday> weekdays;
+  std::set<osmoh::Weekday> weekdays;
   for (int i = 0; i < size; i++)
-    weekdays.insert(ToWeekday(weekdaysArr[i]));
+    weekdays.insert(osmoh::ToWeekday(weekdaysArr[i]));
   tt.SetOpeningDays(weekdays);
   env->ReleaseIntArrayElements(jWeekdays, weekdaysArr, 0);
   tt.SetTwentyFourHours(env->GetBooleanField(jTimetable, g_fidIsFullday));
@@ -167,9 +161,9 @@ TimeTable NativeTimetable(JNIEnv * env, jobject jTimetable)
   return tt;
 }
 
-TimeTableSet NativeTimetableSet(JNIEnv * env, jobjectArray jTimetables)
+editor::ui::TimeTableSet NativeTimetableSet(JNIEnv * env, jobjectArray jTimetables)
 {
-  TimeTableSet tts;
+  editor::ui::TimeTableSet tts;
   int const size = env->GetArrayLength(jTimetables);
   tts.Replace(NativeTimetable(env, env->GetObjectArrayElement(jTimetables, 0)), 0);
 
@@ -234,14 +228,14 @@ JNIEXPORT void Java_app_organicmaps_sdk_editor_OpeningHours_nativeInit(JNIEnv * 
 JNIEXPORT jobjectArray Java_app_organicmaps_sdk_editor_OpeningHours_nativeGetDefaultTimetables(JNIEnv * env,
                                                                                                jclass clazz)
 {
-  TimeTableSet tts;
+  editor::ui::TimeTableSet tts;
   return JavaTimetables(env, tts);
 }
 
 JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeGetComplementTimetable(JNIEnv * env, jclass clazz,
                                                                                             jobjectArray timetables)
 {
-  TimeTableSet const tts = NativeTimetableSet(env, timetables);
+  auto const tts = NativeTimetableSet(env, timetables);
   return JavaTimetable(env, tts.GetComplementTimeTable());
 }
 
@@ -249,9 +243,9 @@ JNIEXPORT jobjectArray Java_app_organicmaps_sdk_editor_OpeningHours_nativeRemove
                                                                                            jobjectArray timetables,
                                                                                            jint ttIndex, jint dayIndex)
 {
-  TimeTableSet tts = NativeTimetableSet(env, timetables);
+  auto tts = NativeTimetableSet(env, timetables);
   auto tt = tts.Get(ttIndex);
-  tt.RemoveWorkingDay(ToWeekday(dayIndex));
+  tt.RemoveWorkingDay(osmoh::ToWeekday(dayIndex));
   tt.Commit();
   return JavaTimetables(env, tts);
 }
@@ -260,9 +254,9 @@ JNIEXPORT jobjectArray Java_app_organicmaps_sdk_editor_OpeningHours_nativeAddWor
                                                                                         jobjectArray timetables,
                                                                                         jint ttIndex, jint dayIndex)
 {
-  TimeTableSet tts = NativeTimetableSet(env, timetables);
+  auto tts = NativeTimetableSet(env, timetables);
   auto tt = tts.Get(ttIndex);
-  tt.AddWorkingDay(ToWeekday(dayIndex));
+  tt.AddWorkingDay(osmoh::ToWeekday(dayIndex));
   tt.Commit();
   return JavaTimetables(env, tts);
 }
@@ -271,7 +265,7 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeSetIsFullda
                                                                                   jobject jTimetable,
                                                                                   jboolean jIsFullday)
 {
-  TimeTable tt = NativeTimetable(env, jTimetable);
+  auto tt = NativeTimetable(env, jTimetable);
   if (jIsFullday)
     tt.SetTwentyFourHours(true);
   else
@@ -286,7 +280,7 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeSetOpeningT
                                                                                     jobject jTimetable,
                                                                                     jobject jOpeningTime)
 {
-  TimeTable tt = NativeTimetable(env, jTimetable);
+  auto tt = NativeTimetable(env, jTimetable);
   tt.SetOpeningTime(NativeTimespan(env, jOpeningTime));
   return JavaTimetable(env, tt);
 }
@@ -295,7 +289,7 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeAddClosedSp
                                                                                    jobject jTimetable,
                                                                                    jobject jClosedSpan)
 {
-  TimeTable tt = NativeTimetable(env, jTimetable);
+  auto tt = NativeTimetable(env, jTimetable);
   tt.AddExcludeTime(NativeTimespan(env, jClosedSpan));
   return JavaTimetable(env, tt);
 }
@@ -304,7 +298,7 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeRemoveClose
                                                                                       jobject jTimetable,
                                                                                       jint jClosedSpanIndex)
 {
-  TimeTable tt = NativeTimetable(env, jTimetable);
+  auto tt = NativeTimetable(env, jTimetable);
   tt.RemoveExcludeTime(static_cast<size_t>(jClosedSpanIndex));
   return JavaTimetable(env, tt);
 }
@@ -313,9 +307,9 @@ JNIEXPORT jobjectArray Java_app_organicmaps_sdk_editor_OpeningHours_nativeTimeta
                                                                                                jclass clazz,
                                                                                                jstring jSource)
 {
-  TimeTableSet tts;
+  editor::ui::TimeTableSet tts;
   std::string const source = jni::ToNativeString(env, jSource);
-  if (!source.empty() && MakeTimeTableSet(OpeningHours(source), tts))
+  if (!source.empty() && editor::MakeTimeTableSet(osmoh::OpeningHours(source), tts))
     return JavaTimetables(env, tts);
 
   return nullptr;
@@ -324,16 +318,16 @@ JNIEXPORT jobjectArray Java_app_organicmaps_sdk_editor_OpeningHours_nativeTimeta
 JNIEXPORT jstring Java_app_organicmaps_sdk_editor_OpeningHours_nativeTimetablesToString(JNIEnv * env, jclass clazz,
                                                                                         jobjectArray jTts)
 {
-  TimeTableSet tts = NativeTimetableSet(env, jTts);
+  auto tts = NativeTimetableSet(env, jTts);
   std::stringstream sstr;
-  sstr << MakeOpeningHours(tts).GetRule();
+  sstr << editor::MakeOpeningHours(tts).GetRule();
   return jni::ToJavaString(env, sstr.str());
 }
 
 JNIEXPORT jboolean Java_app_organicmaps_sdk_editor_OpeningHours_nativeIsTimetableStringValid(JNIEnv * env, jclass clazz,
                                                                                              jstring jSource)
 {
-  return OpeningHours(jni::ToNativeString(env, jSource)).IsValid();
+  return osmoh::OpeningHours(jni::ToNativeString(env, jSource)).IsValid();
 }
 
 JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeGetOpeningHoursInfoFromString(JNIEnv * env,
@@ -342,11 +336,11 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_editor_OpeningHours_nativeGetOpeningH
                                                                                                    jlong jCurrentTime)
 {
   std::string const source = jni::ToNativeString(env, jSource);
-  OpeningHours const oh = OpeningHours(source);
+  osmoh::OpeningHours const oh(source);
 
   if (!source.empty() && oh.IsValid())
   {
-    OpeningHours::InfoT info = oh.GetInfo(static_cast<time_t>(jCurrentTime));
+    osmoh::OpeningHours::InfoT info = oh.GetInfo(static_cast<time_t>(jCurrentTime));
     if (info.state == osmoh::RuleState::Unknown)
       return nullptr;
     return JavaOpeningHoursInfo(env, info.state, oh.IsTwentyFourHours(), info.nextTimeOpen, info.nextTimeClosed);
