+++ android/sdk/src/main/cpp/app/organicmaps/sdk/isolines/IsolinesManager.cpp
@@ -3,7 +3,8 @@
 #include "app/organicmaps/sdk/core/jni_helper.hpp"
 #include "app/organicmaps/sdk/platform/AndroidPlatform.hpp"
 
-using namespace std::placeholders;
+#include <functional>
+#include <memory>
 
 extern "C"
 {
