--- android/sdk/src/main/cpp/app/organicmaps/sdk/routing/LaneInfo.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <jni.h>
-
-#include "routing/lanes/lane_info.hpp"
-
-jobjectArray CreateLanesInfo(JNIEnv * env, routing::turns::lanes::LanesInfo const & lanes);
