--- android/sdk/src/main/cpp/app/organicmaps/sdk/routing/PedestrianDirection.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <jni.h>
-
-#include "app/organicmaps/sdk/core/jni_helper.hpp"
-
-#include "routing/turns.hpp"
-
-jobject ToJavaPedestrianDirection(JNIEnv * env, routing::turns::PedestrianDirection turn);
