--- android/sdk/src/main/cpp/app/organicmaps/sdk/routing/RoutePointInfo.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include "app/organicmaps/sdk/core/jni_helper.hpp"
-
-#include "map/place_page_info.hpp"
-
-jobject CreateRoutePointInfo(JNIEnv * env, place_page::Info const & info)
-{
-  static jclass const clazz = jni::GetGlobalClassRef(env, "app/organicmaps/sdk/routing/RoutePointInfo");
-  static jmethodID const ctorId = jni::GetConstructorID(env, clazz, "(II)V");
-  int const markType = static_cast<int>(info.GetRouteMarkType());
-  return env->NewObject(clazz, ctorId, markType, info.GetIntermediateIndex());
-}
