--- android/sdk/src/main/cpp/app/organicmaps/sdk/routing/RoutingOptions.cpp
@@ -1,37 +1,0 @@
-#include <jni.h>
-#include "app/organicmaps/sdk/Framework.hpp"
-#include "app/organicmaps/sdk/core/jni_helper.hpp"
-#include "routing/routing_options.hpp"
-
-routing::RoutingOptions::Road makeValue(jint option)
-{
-  auto const road = static_cast<uint8_t>(1u << static_cast<int>(option));
-  CHECK_LESS(road, static_cast<uint8_t>(routing::RoutingOptions::Road::Max), ());
-  return static_cast<routing::RoutingOptions::Road>(road);
-}
-
-extern "C"
-{
-JNIEXPORT jboolean Java_app_organicmaps_sdk_routing_RoutingOptions_nativeHasOption(JNIEnv *, jclass, jint option)
-{
-  routing::RoutingOptions routingOptions = routing::RoutingOptions::LoadCarOptionsFromSettings();
-  routing::RoutingOptions::Road road = makeValue(option);
-  return static_cast<jboolean>(routingOptions.Has(road));
-}
-
-JNIEXPORT void Java_app_organicmaps_sdk_routing_RoutingOptions_nativeAddOption(JNIEnv *, jclass, jint option)
-{
-  routing::RoutingOptions routingOptions = routing::RoutingOptions::LoadCarOptionsFromSettings();
-  routing::RoutingOptions::Road road = makeValue(option);
-  routingOptions.Add(road);
-  routing::RoutingOptions::SaveCarOptionsToSettings(routingOptions);
-}
-
-JNIEXPORT void Java_app_organicmaps_sdk_routing_RoutingOptions_nativeRemoveOption(JNIEnv *, jclass, jint option)
-{
-  routing::RoutingOptions routingOptions = routing::RoutingOptions::LoadCarOptionsFromSettings();
-  routing::RoutingOptions::Road road = makeValue(option);
-  routingOptions.Remove(road);
-  routing::RoutingOptions::SaveCarOptionsToSettings(routingOptions);
-}
-}
