--- android/sdk/src/main/cpp/app/organicmaps/sdk/routing/roadshield/RoadShieldInfo.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "app/organicmaps/sdk/core/jni_helper.hpp"
-
-#include "routing/following_info.hpp"
-
-jobject ToJavaRoadShieldInfo(JNIEnv * env, routing::FollowingInfo::RoadShieldInfo const & roadShieldInfo);
