+++ android/sdk/src/main/cpp/app/organicmaps/sdk/sound/tts.cpp
@@ -1,6 +1,7 @@
 #include "app/organicmaps/sdk/Framework.hpp"
 
 #include "app/organicmaps/sdk/core/jni_helper.hpp"
+#include "app/organicmaps/sdk/core/jni_java_methods.hpp"
 
 #include "platform/languages.hpp"
